Implement ref for buildpack config on BuildSettingsTab.
Implement a reference to check if the component is still loading, and to prevent the usage
of a useEffect to update the values over the parent component with an onChange function.

Instead, the component will now receive a reference that will have a function that gets the newest buildConfig object.