add telemetry to preview environment deployment handlers (#3066)
Adding telemetry to preview environment creation handlers so we can start to get a sense for errors that are blocking deployments.

This is an initial introduction of telemetry that captures the request parameters and errors that cause us to exit. If we exit for a reason that is not our fault (like a PR being closed), then I add the attribute unsuccessful-exit-reason to the span. Future commits should extend the telemetry into the functions being called.