fix: use average data correctly on app metric graphs
Previously, when showing "all" data for a given metric, we would show the summed version of the data on the chart. This is confusing to users, as they would assume that the value is the average utilization of whatever the metric was.

Now, for "all" metrics, we show the average instead and remove the old average data line (with min/max still shown separately). Pod-specific charts are still treated normally.