refactor: reduce the number of api queries made when generating the graph
Previously, we would generate up to three queries per chart:

- one for aggregated data
- one for unaggregated data
- one for hpa data (where available)

The unaggregated data was thrown away in favor of the average data, and the queries were fired serially, thus there could be several flashes on the screen before data was shown.

We now instead use a promise to wait for all the requests to happen, and then assign data to the view appropriately. This allows us to "queue" up requests (note: thats not actually what is happening) and avoid queries for unaggregated data where it isn't necessary.