fix: properly query for the cpu/memory autoscaling thresholds
Previously, we would show the sum of all autoscaling thresholds for each pod, resulting in a wildly incorrect graph. Fixing that, we still had our metrics downscaled by half, so we need to decrease the divisor by half in order to get the correct values.

The query itself might be incorrect/simplifyable, but at least for now things work as expected.