refactor: return a simplified list response when retrieving a project list
The project list endpoint doesn't need to respond with all the data. Doing so will mean N calls for feature flags for each project returned, resulting in a n+1 set of queries as we increase the number of feature flags. This would normally be _fine_ but for larger customers, this would make it basically impossible to load the page as the external api calls would potentially time out.