fix: allow fallback to the database for feature flags
The initial implementation of feature flags stored flags in
table-specific columns, making it so we need to use a nasty hack to
fetch the correct value for on-prem deployments. A proper refactor would
be to introduce a migration that moved these flags to a feature flags
table that we could implement a proper WrappedClient for, but a shortcut
is taken here in order to fix this sooner and give us time for a proper
refactor.