feat: manually retrieve the latest version for a given helm release when the version is set to 0
If we use helm's built-in method of retrieving the last release, it will retrieve _all_ releases and then only return the last one. That works ~fine except in cases where you have hundreds of releases, in which case any api calls will take significantly longer.

Instead, we retrieve all non-supersceded versions, then grab the highest version in that list. In the worst case, this would also retrieve every release. In the best case, it only retrieves a single release (the latest).

The ideal case would be if we could sort on the server side, in which case we could Limit results by 1 and not worry about the helm release status label, but Kubernetes only supports limiting results api-side.