fix: properly redirect users to their first project if the project they attempt to route to is non-existent
Due to our use of localStorage for storing the current project, it is possible to get into a state where the current project id is set to a non-existent project, resulting in a weird state for the view that is a melange of all features turned off. This page is hard to get out of because the sidebar doesn't show a project selector, so it appears to the user that there are no projects at all.

Currently, the only known way to trigger such a view is to setup a local dev install, create a new project, browse to the new project, then reset the entire dev cluster. On first login on the new install, the user is redirected to a weird page.

The fix is just to verify the project being viewed is in the list of returned projects. If not, we redirect to the first project in the list.

The ideal case is that we show an error pop up of some sort, but there isn't a standard here, so I've left that error state as is.