refactor: split out the pr workflow into different workflows
This allows developers to skip go or node tests depending on whether they actually changed files that would impact those or not.

Additionally, cancel in-progress tests if a newer commit appears. This will save us time waiting for runners in cases where developers are constantly pushing to the same branch.