fix: persist the pre-deploy tasks when custom helm values are specified
Previously, v1 clusters would drop the pre-deploy job on the floor if there were ever any customizations of the helm values. This is because predeploy tasks are tracked separate from the rest of the umbrella deploy.

This fix parses the entire porter yaml first, and then persists the Release (predeploy) job if there are helm values.