feat: add dependabot for golang dependencies
We have a number of outdated dependencies which make it very difficult to add new packages or update others since we need to find the magic configuration that works with everything.

Instead, lets endeavor to keep everything up to date. A side-effect of this is discovering build issues sooner.

Note that this is limited to two open golang pull requests at this time to avoid flooding the repository.