fix: ensure the injected Dockerfile path is respected for all builds
By default, we'll use the 'Dockerfile' path in the repository. In certain cases though, the specified Dockerfile may be _outside_ of the build context, and thus we inject the Dockerfile from that path into the repository and then perform a build. This change ensures we actually use that Dockerfile vs just writing it into the repo.